Modularise prod
