+++ src/prod/prod_models.rs
@@ -1,0 +1,1 @@
+pub struct Prod;
