+++ src/prod/prod_user_dirs.rs
@@ -1,0 +1,17 @@
+use crate::models::HopEffect;
+use crate::algebra::user_dirs::UserDirs;
+use super::prod_models::Prod;
+use dirs::home_dir;
+
+use std::io;
+use std::path::PathBuf;
+
+impl UserDirs for Prod {
+  fn get_hop_home(&self, path: &str) -> HopEffect<PathBuf> {
+    Ok(get_home()?.join(path))
+  }
+}
+
+fn get_home() -> HopEffect<PathBuf> {
+    home_dir().ok_or_else(|| io::Error::new(io::ErrorKind::Other, "Could not get home directory"))
+}
